Change ArticleDao ID type from Long to Int

+++ src/main/kotlin/io/github/gunkim/realworld/infrastructure/jdbc/article/dao/ArticleDao.kt
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 
-interface ArticleDao : JpaRepository<ArticleJpaEntity, Long> {
+interface ArticleDao : JpaRepository<ArticleJpaEntity, Int> {
 
     /**
      * TODO: Consider using a query builder for optional conditions.
